refactor(core): extract view highlighter
